improve the book lookup
Filter by series name, first look for an exact match and if none are
found then look for a prefix.
The comparison is not case-sensitive and punctuation is ignored (only '!'
for now, will enrich as needed).

With this improved logic, we no longer need to filter on Asian
productions (some mangas are written by Westerners).