use Searx instead of bedetheque search
Bethedeque search is very basic, and doesn't deal with variations.
Trying manually to please it by "normalizing" the names is an arm race
and is quickly getting really complex/with a lot of special cases.

There is a class of software that handle this kind of problematic really
well: search engine.
And most of them can be told to limit the search to a dedicated website
so we could use that to find the book page on bedetheque.

Now, mainstream search engine seems to be pretty agressive on rate
limitation/blocking scrapper, even DuckDuckGo is giving a lot of 403 :(

Fortuntely, there is this great project of an open-source distributed
meta search engine: Searx

For now, we use a single instance and limit ourselve to ~3 query/min,
but it can be improved later.

We search on different Searx instances to avoid spamming a single
instance (so we can reduce the delay between each call) and be resilient
(if a given instance is down we will try another).

The serverlist is retrieved at startup and filtered (we remove offline
server, and blacklist the one that doesn't seem to know about
bedetheque).